+++ changelog.md
@@ -9,6 +9,8 @@
  - Shift+Click Column Header : Sorts the list, but will keep the selected items grouped to the top of the list.
  - Shift+Click Reset Button : Randomizes the order of the list.
  - Added 'Difference (Side)' stereo -> mono processing option.
+ - Trim All Selected action.
+ - Merge All Selected action - mixes down all the selected items into one sample.
 
 2023-04-07 1.3.1
  - Fixed bug where a duplicate sample shared the sources audio-buffer in specific situations (e.g. when reversing the sample.)